updated orders insert view order and cart insert

+++ src/main/java/com/pojo/UpdateWalletPojo.java
@@ -3,7 +3,7 @@ package com.pojo;
 public class UpdateWalletPojo {
 	private int userId;
 	private String password;
-	private double price;
+	private long price;
 	public int getUserId() {
 		return userId;
 	}
@@ -16,22 +16,20 @@ public class UpdateWalletPojo {
 	public void setPassword(String password) {
 		this.password = password;
 	}
-	
-	public double getPrice() {
+	public long getPrice() {
 		return price;
 	}
-	public void setPrice(double price) {
+	public void setPrice(long price) {
 		this.price = price;
 	}
 	
-	public UpdateWalletPojo(int userId, String password, double price) {
+	
+	public UpdateWalletPojo(int userId, String password, long price) {
 		super();
 		this.userId = userId;
 		this.password = password;
 		this.price = price;
 	}
-	
-	
 	@Override
 	public String toString() {
 		return "UpdateWalletPojo [userId=" + userId + ", password=" + password + ", price=" + price + "]";
